Split SystemToSchedule to a separate file
