+++ Thrive.csproj
@@ -31,6 +31,7 @@
     <Compile Include="src\native\interop\CPUCheckResult.cs" />
     <Compile Include="src\saving\serializers\ConditionSetConverter.cs" />
     <Compile Include="src\tools\GenerateThreadedSystems.cs" />
+    <Compile Include="src\tools\SystemToSchedule.cs" />
     <Compile Include="src\tutorial\microbe_editor\AtpBalanceIntroduction.cs" />
     <Compile Include="src\tutorial\microbe_editor\ChemoreceptorPlacementTutorial.cs" />
     <Compile Include="src\tutorial\microbe_editor\FlagellumPlacementTutorial.cs" />
