+++ src/tools/GenerateThreadedSystems.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
 using DefaultEcs.System;
 using Godot;
@@ -28,27 +27,26 @@ public class GenerateThreadedSystems : Node
         };
 
     private readonly Type systemBaseType = typeof(ISystem<>);
-    private readonly Type systemWithAttribute = typeof(WithAttribute);
 
     // private readonly Type systemWithoutAttribute = typeof(WithoutAttribute);
-
-    private readonly Type writesToAttribute = typeof(WritesToComponentAttribute);
-    private readonly Type readsFromAttribute = typeof(ReadsComponentAttribute);
-    private readonly Type runsAfterAttribute = typeof(RunsAfterAttribute);
-    private readonly Type runsBeforeAttribute = typeof(RunsBeforeAttribute);
-    private readonly Type runsOnMainAttribute = typeof(RunsOnMainThreadAttribute);
-    private readonly Type conditionalRunAttribute = typeof(RunsConditionallyAttribute);
-    private readonly Type customRunAttribute = typeof(RunsWithCustomCodeAttribute);
-
-    private readonly Type runsOnFrameAttribute = typeof(RunsOnFrameAttribute);
-
-    private readonly Type compilerGeneratedAttribute = typeof(CompilerGeneratedAttribute);
 
     private bool done;
 
     private GenerateThreadedSystems()
     {
         random = new Random(DefaultSeed);
+    }
+
+    public static void EnsureOneBlankLine(List<string> lines, bool acceptBlockStart = true)
+    {
+        if (lines.Count < 1)
+            return;
+
+        var line = lines[lines.Count - 1];
+        if (!string.IsNullOrWhiteSpace(line) && (!acceptBlockStart || !line.EndsWith("{")))
+        {
+            lines.Add(string.Empty);
+        }
     }
 
     public override void _Ready()
@@ -77,45 +75,6 @@ public class GenerateThreadedSystems : Node
         finally
         {
             done = true;
-        }
-    }
-
-    private static string GetIndentForText(int indent)
-    {
-        if (indent < 1)
-            return string.Empty;
-
-        return new string(' ', indent);
-    }
-
-    private static int DetectLineIndentationLevel(string line)
-    {
-        int spaceCount = 0;
-
-        foreach (var character in line)
-        {
-            if (character <= ' ')
-            {
-                ++spaceCount;
-            }
-            else
-            {
-                break;
-            }
-        }
-
-        return spaceCount;
-    }
-
-    private static void EnsureOneBlankLine(List<string> lines, bool acceptBlockStart = true)
-    {
-        if (lines.Count < 1)
-            return;
-
-        var line = lines[lines.Count - 1];
-        if (!string.IsNullOrWhiteSpace(line) && (!acceptBlockStart || !line.EndsWith("{")))
-        {
-            lines.Add(string.Empty);
         }
     }
 
@@ -183,7 +142,7 @@ public class GenerateThreadedSystems : Node
         var groups = new List<ExecutionGroup>();
 
         // Add systems that are considered equal to the same execution group
-        var comparer = new SystemRequirementsBasedComparer();
+        var comparer = new SystemToSchedule.SystemRequirementsBasedComparer();
 
         foreach (var mainSystem in mainSystems)
         {
@@ -260,17 +219,106 @@ public class GenerateThreadedSystems : Node
         bool nextMainShouldBlock = false;
         bool skipOtherBlock = false;
 
+        bool hadOtherBlock = false;
+
         foreach (var group in groups)
         {
             // Probably fine to have inter-group dependencies when adding the blocks
             // mainReads.Clear();
             // mainWrites.Clear();
+            // seenMainSystems
 
             using var mainEnumerator = group.Systems.Where(s => s.RunsOnMainThread).GetEnumerator();
             using var otherEnumerator = group.Systems.Where(s => !s.RunsOnMainThread).GetEnumerator();
 
             bool firstMain = true;
             bool firstOther = true;
+
+            /*foreach (var mainSystem in group.Systems.Where(s => s.RunsOnMainThread))
+            {
+                // Require barrier on first group item (but not required on the first group)
+                if (firstMain && !firstMainGroup)
+                {
+                    firstMain = false;
+
+                    // Also don't require barrier if there hasn't been another thread in between the main threads
+                    if (hadOtherBlock)
+                    {
+                        mainSystem.RequiresBarrierBefore = true;
+                        hadOtherBlock = false;
+                    }
+                }
+
+                foreach (var usedComponent in mainSystem.WritesComponents)
+                {
+                    mainWrites.Add(usedComponent);
+                }
+
+                foreach (var usedComponent in mainSystem.ReadsComponents)
+                {
+                    mainReads.Add(usedComponent);
+                }
+
+                seenMainSystems.Add(mainSystem);
+
+                if (nextMainShouldBlock)
+                {
+                    nextMainShouldBlock = false;
+
+                    // Ensure consistent number of thread waits
+                    if (!mainSystem.RequiresBarrierBefore)
+                    {
+                        mainSystem.RequiresBarrierBefore = true;
+                    }
+                    else
+                    {
+                        skipOtherBlock = true;
+                    }
+                }
+
+                firstMainGroup = false;
+            }
+
+            foreach (var otherSystem in group.Systems.Where(s => !s.RunsOnMainThread))
+            {
+                if (firstOther && !firstOtherGroup)
+                {
+                    firstOther = false;
+                    if (!skipOtherBlock)
+                    {
+                        if (otherSystem.RequiresBarrierBefore)
+                            throw new Exception("This shouldn't be able to be true here");
+
+                        otherSystem.RequiresBarrierBefore = true;
+                    }
+                    else
+                    {
+                        skipOtherBlock = false;
+                    }
+                }
+
+                // If conflicts with a main system needs to block, and main needs to wait after the current main
+                // system has ran
+                if (otherSystem.WritesComponents.Any(c =>
+                        mainReads.Contains(c) || mainWrites.Contains(c)) ||
+                    otherSystem.ReadsComponents.Any(c => mainWrites.Contains(c)) ||
+                    seenMainSystems.Any(s => comparer.CompareWeak(otherSystem, s) < 0))
+                {
+                    if (!otherSystem.RequiresBarrierBefore && !nextMainShouldBlock)
+                    {
+                        otherSystem.RequiresBarrierBefore = true;
+                        nextMainShouldBlock = true;
+
+                        // Clear the interference info as we've decided on a blocking location
+                        mainReads.Clear();
+                        mainWrites.Clear();
+                        seenMainSystems.Clear();
+                    }
+                }
+
+                firstOtherGroup = false;
+                hadOtherBlock = true;
+            }*/
 
             while (true)
             {
@@ -282,7 +330,13 @@ public class GenerateThreadedSystems : Node
                     if (firstMain && !firstMainGroup)
                     {
                         firstMain = false;
-                        mainEnumerator.Current.RequiresBarrierBefore = true;
+
+                        // Also don't require barrier if there hasn't been another thread in between the main threads
+                        if (hadOtherBlock)
+                        {
+                            mainEnumerator.Current.RequiresBarrierBefore = true;
+                            hadOtherBlock = false;
+                        }
                     }
 
                     foreach (var usedComponent in mainEnumerator.Current.WritesComponents)
@@ -322,6 +376,9 @@ public class GenerateThreadedSystems : Node
                         firstOther = false;
                         if (!skipOtherBlock)
                         {
+                            if (otherEnumerator.Current.RequiresBarrierBefore)
+                                throw new Exception("This shouldn't be able to be true here");
+
                             otherEnumerator.Current.RequiresBarrierBefore = true;
                         }
                         else
@@ -337,13 +394,16 @@ public class GenerateThreadedSystems : Node
                         otherEnumerator.Current.ReadsComponents.Any(c => mainWrites.Contains(c)) ||
                         seenMainSystems.Any(s => comparer.CompareWeak(otherEnumerator.Current, s) < 0))
                     {
-                        otherEnumerator.Current.RequiresBarrierBefore = true;
-                        nextMainShouldBlock = true;
+                        if (!otherEnumerator.Current.RequiresBarrierBefore && !nextMainShouldBlock)
+                        {
+                            otherEnumerator.Current.RequiresBarrierBefore = true;
+                            nextMainShouldBlock = true;
 
-                        // Clear the interference info as we've decided on a blocking location
-                        mainReads.Clear();
-                        mainWrites.Clear();
-                        seenMainSystems.Clear();
+                            // Clear the interference info as we've decided on a blocking location
+                            mainReads.Clear();
+                            mainWrites.Clear();
+                            seenMainSystems.Clear();
+                        }
                     }
 
                     seenAnything = true;
@@ -360,8 +420,37 @@ public class GenerateThreadedSystems : Node
                 firstOtherGroup = false;
         }
 
+        CheckBarrierCounts(groups);
+
         // Generate the final results
         WriteResultOfThreadedRunning(groups, processSystemTextLines, variables);
+    }
+
+    private void CheckBarrierCounts(List<ExecutionGroup> groups)
+    {
+        int mainBarriers = 0;
+        int otherBarriers = 0;
+
+        foreach (var group in groups)
+        {
+            foreach (var system in group.Systems)
+            {
+                if (system.RequiresBarrierBefore)
+                {
+                    if (system.RunsOnMainThread)
+                    {
+                        ++mainBarriers;
+                    }
+                    else
+                    {
+                        ++otherBarriers;
+                    }
+                }
+            }
+        }
+
+        if (mainBarriers != otherBarriers)
+            throw new Exception("Uneven barrier counts, run will get blocked");
     }
 
     private void WriteResultOfThreadedRunning(List<ExecutionGroup> groups, List<string> lineReceiver,
@@ -372,7 +461,7 @@ public class GenerateThreadedSystems : Node
 
         // Task for background operations
         lineReceiver.Add("var background1 = new Task(() =>");
-        lineReceiver.Add($"{GetIndentForText(4)}{{");
+        lineReceiver.Add($"{StringUtils.GetIndent(4)}{{");
 
         foreach (var group in groups)
         {
@@ -380,9 +469,9 @@ public class GenerateThreadedSystems : Node
         }
 
         lineReceiver.Add(string.Empty);
-        lineReceiver.Add($"{GetIndentForText(4)}barrier1.SignalAndWait();");
+        lineReceiver.Add($"{StringUtils.GetIndent(8)}barrier1.SignalAndWait();");
 
-        lineReceiver.Add($"{GetIndentForText(4)}}});");
+        lineReceiver.Add($"{StringUtils.GetIndent(4)}}});");
 
         lineReceiver.Add(string.Empty);
         lineReceiver.Add("TaskExecutor.Instance.AddTask(background1);");
@@ -421,7 +510,7 @@ public class GenerateThreadedSystems : Node
     /// <returns>The value in <see cref="systems"/></returns>
     private IList<SystemToSchedule> SortSingleGroupOfSystems(IList<SystemToSchedule> systems)
     {
-        var comparer = new SystemRequirementsBasedComparer();
+        var comparer = new SystemToSchedule.SystemRequirementsBasedComparer();
 
         // This is really not efficient but for now the sorting needs to continue until all the constraints are taken
         // into account and nothing is sorted anymore
@@ -492,7 +581,7 @@ public class GenerateThreadedSystems : Node
 
     private void VerifyOrderOfSystems(IList<SystemToSchedule> systems)
     {
-        var comparer = new SystemRequirementsBasedComparer();
+        var comparer = new SystemToSchedule.SystemRequirementsBasedComparer();
 
         for (int i = 0; i < systems.Count; ++i)
         {
@@ -520,18 +609,16 @@ public class GenerateThreadedSystems : Node
     {
         var result = new List<SystemToSchedule>();
 
-        int order = 0;
-
         foreach (var field in type.GetFields(BindingFlags.Public | BindingFlags.Instance | BindingFlags.NonPublic))
         {
             // Skip automatic property backing fields
-            if (field.CustomAttributes.Any(a => a.AttributeType == compilerGeneratedAttribute))
+            if (field.CustomAttributes.Any(a => a.AttributeType == SystemToSchedule.CompilerGeneratedAttribute))
                 continue;
 
             if (!field.FieldType.IsAssignableToGenericType(systemBaseType))
                 continue;
 
-            result.Add(new SystemToSchedule(field.FieldType, field.Name, order++));
+            result.Add(new SystemToSchedule(field.FieldType, field.Name));
         }
 
         foreach (var property in type.GetProperties(
@@ -544,22 +631,22 @@ public class GenerateThreadedSystems : Node
             if (result.Any(r => r.FieldName.Equals(property.Name, StringComparison.OrdinalIgnoreCase)))
                 continue;
 
-            result.Add(new SystemToSchedule(property.PropertyType, property.Name, order++));
+            result.Add(new SystemToSchedule(property.PropertyType, property.Name));
         }
 
         foreach (var systemToSchedule in result)
         {
-            ParseSystemAttributes(systemToSchedule);
+            SystemToSchedule.ParseSystemAttributes(systemToSchedule);
         }
 
-        ResolveSystemDependencies(result);
+        SystemToSchedule.ResolveSystemDependencies(result);
 
         // Sanity check no duplicate systems found
         if (result.GroupBy(s => s.Type).Any(g => g.Count() > 1))
             throw new Exception("Some type of system is included multiple times");
 
         // Make sure sorting works sensibly
-        var comparer = new SystemRequirementsBasedComparer();
+        var comparer = new SystemToSchedule.SystemRequirementsBasedComparer();
         foreach (var item1 in result)
         {
             foreach (var item2 in result)
@@ -581,233 +668,6 @@ public class GenerateThreadedSystems : Node
         }
 
         return result;
-    }
-
-    private void ParseSystemAttributes(SystemToSchedule systemToSchedule)
-    {
-        systemToSchedule.RunsOnFrame = systemToSchedule.Type.GetCustomAttribute(runsOnFrameAttribute) != null;
-        systemToSchedule.RunsOnMainThread = systemToSchedule.Type.GetCustomAttribute(runsOnMainAttribute) != null;
-
-        var conditionalRun = systemToSchedule.Type.GetCustomAttribute(conditionalRunAttribute);
-
-        if (conditionalRun != null)
-        {
-            systemToSchedule.RunCondition = ((RunsConditionallyAttribute)conditionalRun).Condition;
-        }
-
-        var customRun = systemToSchedule.Type.GetCustomAttribute(customRunAttribute);
-
-        if (customRun != null)
-        {
-            systemToSchedule.CustomRunCode = ((RunsWithCustomCodeAttribute)customRun).CustomCode;
-        }
-
-        var expectedWritesTo = new List<Type>();
-
-        var withRaw = systemToSchedule.Type.GetCustomAttributes(systemWithAttribute);
-
-        foreach (var attributeRaw in withRaw)
-        {
-            var attribute = (WithAttribute)attributeRaw;
-
-            // TODO: handle without attributes?
-            if (attribute.FilterType is not (ComponentFilterType.WithoutEither or ComponentFilterType.Without))
-            {
-                foreach (var componentType in attribute.ComponentTypes)
-                {
-                    expectedWritesTo.Add(componentType);
-                }
-            }
-        }
-
-        var expectedReadsFrom = new List<Type>();
-
-        var readsRaw = systemToSchedule.Type.GetCustomAttributes(readsFromAttribute);
-
-        foreach (var attributeRaw in readsRaw)
-        {
-            var attribute = (ReadsComponentAttribute)attributeRaw;
-
-            // Convert writes to reads if exist
-            expectedWritesTo.Remove(attribute.ReadsFrom);
-
-            if (!expectedReadsFrom.Contains(attribute.ReadsFrom))
-                expectedReadsFrom.Add(attribute.ReadsFrom);
-        }
-
-        var writesRaw = systemToSchedule.Type.GetCustomAttributes(writesToAttribute);
-
-        foreach (var attributeRaw in writesRaw)
-        {
-            var attribute = (WritesToComponentAttribute)attributeRaw;
-
-            if (expectedReadsFrom.Contains(attribute.WritesTo))
-            {
-                throw new Exception(
-                    "Shouldn't specify a writes to component with already a read attribute for same type");
-            }
-
-            if (!expectedWritesTo.Contains(attribute.WritesTo))
-                expectedWritesTo.Add(attribute.WritesTo);
-        }
-
-        // TODO: should the following be done?:
-        // All writes are also reads for simplicity in checking thread access
-
-        systemToSchedule.WritesComponents = expectedWritesTo;
-        systemToSchedule.ReadsComponents = expectedReadsFrom;
-    }
-
-    private void ResolveSystemDependencies(List<SystemToSchedule> systems)
-    {
-        SystemToSchedule GetSystemByType(Type type)
-        {
-            return systems.First(s => s.Type == type);
-        }
-
-        foreach (var system in systems)
-        {
-            var runsBeforeRaw = system.Type.GetCustomAttributes(runsBeforeAttribute);
-
-            foreach (var attributeRaw in runsBeforeRaw)
-            {
-                var attribute = (RunsBeforeAttribute)attributeRaw;
-
-                var otherSystem = GetSystemByType(attribute.BeforeSystem);
-                system.RunsBefore.Add(otherSystem);
-
-                // TODO: should we instead require explicit attributes on both sides of the relationship?
-                // That would require more lines of code but there would be less hidden things to know about when
-                // reading a class.
-
-                // Add the other side of the relationship automatically
-                otherSystem.RunsAfter.Add(system);
-            }
-
-            var runsAfterRaw = system.Type.GetCustomAttributes(runsAfterAttribute);
-
-            foreach (var attributeRaw in runsAfterRaw)
-            {
-                var attribute = (RunsAfterAttribute)attributeRaw;
-
-                var otherSystem = GetSystemByType(attribute.AfterSystem);
-                system.RunsAfter.Add(otherSystem);
-                otherSystem.RunsBefore.Add(system);
-            }
-        }
-
-        // Detect cycles in runs before or runs after
-        // TODO: this is likely not good enough currently
-        var seenSystems = new HashSet<SystemToSchedule>();
-
-        foreach (var system in systems)
-        {
-            if (system.RunsAfter.Any(s => system.RunsBefore.Contains(s)))
-                throw new Exception("A system is set to run both after and before another");
-
-            if (system.RunsBefore.Any(s => system.RunsAfter.Contains(s)))
-                throw new Exception("A system is set to run both after and before another");
-
-            // System dependencies are not allowed to run after itself
-            DoNotAllowSeeingRunsAfter(system, system, seenSystems);
-            seenSystems.Clear();
-
-            // Or before itself
-            DoNotAllowSeeingRunsBefore(system, system, seenSystems);
-            seenSystems.Clear();
-        }
-
-        // Add recursive dependencies
-        foreach (var system in systems)
-        {
-            foreach (var runsAfter in system.RunsAfter.ToList())
-            {
-                CollectRunsAfter(runsAfter, system.RunsAfter);
-            }
-
-            if (system.RunsAfter.Contains(system))
-                throw new Exception("System ended up running after itself after recursive resolve");
-
-            foreach (var runsBefore in system.RunsBefore.ToList())
-            {
-                CollectRunsBefore(runsBefore, system.RunsBefore);
-            }
-
-            if (system.RunsBefore.Contains(system))
-                throw new Exception("System ended up running before itself after recursive resolve");
-        }
-    }
-
-    private void DoNotAllowSeeingRunsAfter(SystemToSchedule systemToNotRunAfter,
-        SystemToSchedule checkFrom, HashSet<SystemToSchedule> alreadyVisited)
-    {
-        foreach (var systemToSchedule in checkFrom.RunsAfter)
-        {
-            if (!alreadyVisited.Add(systemToSchedule))
-                continue;
-
-            // Check if we found a dependency on a system that is not allowed (but only if processing a list of the
-            // wanted type, otherwise this is just meant to recursively traverse to find more lists of the right type)
-            if (systemToSchedule == systemToNotRunAfter && !systemToNotRunAfter.RunsBefore.Contains(checkFrom))
-            {
-                throw new Exception("Seen a system reference a system it shouldn't run after " +
-                    $"({systemToNotRunAfter.Type.Name} depends on {checkFrom.Type.Name})");
-            }
-
-            DoNotAllowSeeingRunsAfter(systemToNotRunAfter, systemToSchedule, alreadyVisited);
-        }
-
-        foreach (var systemToSchedule in checkFrom.RunsBefore)
-        {
-            if (!alreadyVisited.Add(systemToSchedule))
-                continue;
-
-            DoNotAllowSeeingRunsAfter(systemToNotRunAfter, systemToSchedule, alreadyVisited);
-        }
-    }
-
-    private void DoNotAllowSeeingRunsBefore(SystemToSchedule systemToNotRunBefore,
-        SystemToSchedule checkFrom, HashSet<SystemToSchedule> alreadyVisited)
-    {
-        foreach (var systemToSchedule in checkFrom.RunsAfter)
-        {
-            if (!alreadyVisited.Add(systemToSchedule))
-                continue;
-
-            DoNotAllowSeeingRunsBefore(systemToNotRunBefore, systemToSchedule, alreadyVisited);
-        }
-
-        foreach (var systemToSchedule in checkFrom.RunsBefore)
-        {
-            if (!alreadyVisited.Add(systemToSchedule))
-                continue;
-
-            if (systemToSchedule == systemToNotRunBefore && !systemToNotRunBefore.RunsAfter.Contains(checkFrom))
-            {
-                throw new Exception("Seen a system reference a system it shouldn't run after " +
-                    $"({systemToNotRunBefore.Type.Name} depends on {checkFrom.Type.Name})");
-            }
-
-            DoNotAllowSeeingRunsBefore(systemToNotRunBefore, systemToSchedule, alreadyVisited);
-        }
-    }
-
-    private void CollectRunsAfter(SystemToSchedule systemToStart, HashSet<SystemToSchedule> result)
-    {
-        foreach (var runsAfter in systemToStart.RunsAfter)
-        {
-            if (result.Add(runsAfter))
-                CollectRunsAfter(runsAfter, result);
-        }
-    }
-
-    private void CollectRunsBefore(SystemToSchedule systemToStart, HashSet<SystemToSchedule> result)
-    {
-        foreach (var runsBefore in systemToStart.RunsBefore)
-        {
-            if (result.Add(runsBefore))
-                CollectRunsBefore(runsBefore, result);
-        }
     }
 
     private (List<SystemToSchedule> MainSystems, List<SystemToSchedule> OtherSystems) SplitSystemsToMainThread(
@@ -852,6 +712,7 @@ public class GenerateThreadedSystems : Node
         writer.WriteLine('{');
 
         indent += 4;
+        bool addedVariables = false;
 
         foreach (var pair in variables)
         {
@@ -859,32 +720,37 @@ public class GenerateThreadedSystems : Node
 
             var initializer = info.Initializer != null ? $" = {info.Initializer}" : string.Empty;
 
-            writer.WriteLine(GetIndentForText(indent) +
+            writer.WriteLine(StringUtils.GetIndent(indent) +
                 $"private {(info.IsReadonly ? "readonly " : string.Empty)}{info.Type} {pair.Key}{initializer};");
+
+            addedVariables = true;
         }
 
-        writer.WriteLine(GetIndentForText(indent) + "protected override void OnProcessFixedLogic(float delta)");
+        if (addedVariables)
+            writer.WriteLine();
+
+        writer.WriteLine(StringUtils.GetIndent(indent) + "protected override void OnProcessFixedLogic(float delta)");
         indent = WriteBlockContents(writer, process, indent);
 
         writer.WriteLine();
-        writer.WriteLine(GetIndentForText(indent) + "private void OnProcessFrameLogic(float delta)");
+        writer.WriteLine(StringUtils.GetIndent(indent) + "private void OnProcessFrameLogic(float delta)");
         indent = WriteBlockContents(writer, processFrame, indent);
 
         writer.WriteLine();
-        writer.WriteLine(GetIndentForText(indent) + "private void DisposeGenerated()");
-        writer.WriteLine(GetIndentForText(indent) + "{");
+        writer.WriteLine(StringUtils.GetIndent(indent) + "private void DisposeGenerated()");
+        writer.WriteLine(StringUtils.GetIndent(indent) + "{");
         indent += 4;
 
         foreach (var pair in variables)
         {
             if (pair.Value.Dispose)
             {
-                writer.WriteLine(GetIndentForText(indent) + $"{pair.Key}.Dispose();");
+                writer.WriteLine(StringUtils.GetIndent(indent) + $"{pair.Key}.Dispose();");
             }
         }
 
         indent -= 4;
-        writer.WriteLine(GetIndentForText(indent) + "}");
+        writer.WriteLine(StringUtils.GetIndent(indent) + "}");
 
         // End of class
         writer.WriteLine('}');
@@ -896,7 +762,7 @@ public class GenerateThreadedSystems : Node
 
     private int WriteBlockContents(StreamWriter writer, List<string> lines, int indent)
     {
-        writer.WriteLine(GetIndentForText(indent) + "{");
+        writer.WriteLine(StringUtils.GetIndent(indent) + "{");
         indent += 4;
 
         foreach (var line in lines)
@@ -907,112 +773,12 @@ public class GenerateThreadedSystems : Node
                 continue;
             }
 
-            writer.WriteLine(GetIndentForText(indent) + line);
+            writer.WriteLine(StringUtils.GetIndent(indent) + line);
         }
 
         indent -= 4;
-        writer.WriteLine(GetIndentForText(indent) + "}");
+        writer.WriteLine(StringUtils.GetIndent(indent) + "}");
         return indent;
-    }
-
-    private class SystemToSchedule
-    {
-        public readonly Type Type;
-        public readonly string FieldName;
-        public readonly int OriginalOrder;
-
-        public string? RunCondition;
-        public string? CustomRunCode;
-
-        public bool RunsOnFrame;
-        public bool RunsOnMainThread;
-
-        public HashSet<SystemToSchedule> RunsBefore = new();
-        public HashSet<SystemToSchedule> RunsAfter = new();
-
-        public List<Type> ReadsComponents = new();
-        public List<Type> WritesComponents = new();
-
-        public bool RequiresBarrierBefore;
-
-        public SystemToSchedule(Type type, string name, int order)
-        {
-            Type = type;
-            FieldName = name;
-            OriginalOrder = order;
-        }
-
-        public bool ShouldRunBefore(SystemToSchedule other)
-        {
-            // This depends on the data setup to already check for conflicting (circular) run relationships
-            if (RunsBefore.Contains(other))
-                return true;
-
-            if (other.RunsAfter.Contains(this))
-                return true;
-
-            return false;
-        }
-
-        public bool WantsToWriteBefore(SystemToSchedule other)
-        {
-            return WritesComponents.Any(c => other.ReadsComponents.Contains(c) && !other.WritesComponents.Contains(c));
-        }
-
-        public bool ShouldRunAfter(SystemToSchedule other)
-        {
-            if (RunsAfter.Contains(other))
-                return true;
-
-            if (other.RunsBefore.Contains(this))
-                return true;
-
-            return false;
-        }
-
-        public void GetRunningText(List<string> lineReceiver, int indent)
-        {
-            if (RequiresBarrierBefore)
-            {
-                lineReceiver.Add(GetIndentForText(indent) + "barrier1.SignalAndWait();");
-            }
-
-            bool closeBrace = false;
-
-            if (RunCondition != null)
-            {
-                lineReceiver.Add(GetIndentForText(indent) + $"if ({RunCondition})");
-                lineReceiver.Add(GetIndentForText(indent) + '{');
-
-                // Indent inside the condition
-                indent += 4;
-                closeBrace = true;
-            }
-
-            if (CustomRunCode != null)
-            {
-                foreach (var customLine in CustomRunCode.Split("\n"))
-                {
-                    lineReceiver.Add(GetIndentForText(indent) + string.Format(customLine, FieldName));
-                }
-            }
-            else
-            {
-                lineReceiver.Add(GetIndentForText(indent) + $"{FieldName}.Update(delta);");
-            }
-
-            if (closeBrace)
-            {
-                indent -= 4;
-                lineReceiver.Add(GetIndentForText(indent) + '}');
-                EnsureOneBlankLine(lineReceiver);
-            }
-        }
-
-        public override string ToString()
-        {
-            return Type.Name;
-        }
     }
 
     private class ExecutionGroup
@@ -1043,7 +809,7 @@ public class GenerateThreadedSystems : Node
                 {
                     first = false;
                     EnsureOneBlankLine(lineReceiver);
-                    lineReceiver.Add(GetIndentForText(indent) + comment);
+                    lineReceiver.Add(StringUtils.GetIndent(indent) + comment);
                 }
 
                 system.GetRunningText(lineReceiver, indent);
@@ -1063,74 +829,6 @@ public class GenerateThreadedSystems : Node
             Type = type;
             Initializer = initializer;
             IsReadonly = isReadonly;
-        }
-    }
-
-    // NOTE: this doesn't work for a single pass sort as this can consider next to each other items equal, but then
-    // items on either side of that block of non-sortables need to be sorted around the block
-    private class SystemRequirementsBasedComparer : IComparer<SystemToSchedule>
-    {
-        public int Compare(SystemToSchedule x, SystemToSchedule y)
-        {
-            if (ReferenceEquals(x, y))
-                return 0;
-            if (ReferenceEquals(null, y))
-                return 1;
-            if (ReferenceEquals(null, x))
-                return -1;
-
-            if (x.ShouldRunBefore(y))
-                return -1;
-
-            if (y.ShouldRunBefore(x))
-                return 1;
-
-            if (x.ShouldRunAfter(y))
-                return 1;
-
-            if (y.ShouldRunAfter(x))
-                return -1;
-
-            // Writes before reads ordering, but only if clean order can be found. Both ways need to be checked as
-            // there are systems that both write to a component type that the other just reads.
-            // Might be fun to debug print info when systems cannot do this cleanly to give potential places to improve
-            // the systems decoupling
-            bool xWantsWrite = x.WantsToWriteBefore(y);
-            bool yWantsWrite = y.WantsToWriteBefore(x);
-            if (xWantsWrite && !yWantsWrite)
-                return -1;
-
-            if (yWantsWrite && !xWantsWrite)
-                return 1;
-
-            // No requirement for either system to run before the other
-            return 0;
-        }
-
-        public int CompareWeak(SystemToSchedule x, SystemToSchedule y)
-        {
-            if (ReferenceEquals(x, y))
-                return 0;
-            if (ReferenceEquals(null, y))
-                return 1;
-            if (ReferenceEquals(null, x))
-                return -1;
-
-            if (x.ShouldRunBefore(y))
-                return -1;
-
-            if (y.ShouldRunBefore(x))
-                return 1;
-
-            if (x.ShouldRunAfter(y))
-                return 1;
-
-            if (y.ShouldRunAfter(x))
-                return -1;
-
-            // Weak variant that doesn't enforce write / read relationships
-
-            return 0;
         }
     }
 }
